Refactor QuantityDispensed control
Control now takes the dispensed quantity as input
