+++ client/packages/common/src/intl/locales/en/programs.json
@@ -1,5 +1,6 @@
 {
   "control.adherence-status-warning": "Pill Count: after last visit = {{previousCountOnHand}} < before visit = {{remainingCount}}",
   "control.note.text-label": "Text",
-  "control.note.author-label": "Author"
+  "control.note.author-label": "Author",
+  "label.total-quantity": "Total Quantity"
 }
