+++ client/packages/programs/src/JsonForms/components/index.ts
@@ -5,5 +5,5 @@ export * from './DecisionTree/DecisionTree';
 export * from './EncounterLineChart';
 export * from './IdGenerator';
 export * from './PreviousEncounterField';
-export * from './QuantityPrescribed';
+export * from './QuantityDispensed';
 export * from './EncounterProgramEvent';
